requirements.txt
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via pytest
django==4.1.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-model-utils
    #   django-notifications-hq
    #   djangorestframework
    #   jsonfield
django-cors-headers==3.13.0
    # via -r requirements.in
django-model-utils==4.3.1
    # via django-notifications-hq
django-notifications-hq==1.7.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
exceptiongroup==1.1.0
    # via pytest
flake8==6.0.0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
jsonfield==3.1.0
    # via django-notifications-hq
mccabe==0.7.0
    # via flake8
mypy==0.991
    # via -r requirements.in
mypy-extensions==0.4.3
    # via mypy
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via pytest
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pytest==7.2.1
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.in
pytz==2022.7.1
    # via
    #   django-notifications-hq
    #   djangorestframework
pyyaml==6.0
    # via -r requirements.in
sqlparse==0.4.3
    # via django
swapper==1.3.0
    # via django-notifications-hq
tomli==2.0.1
    # via
    #   mypy
    #   pytest
typing-extensions==4.4.0
    # via mypy
